Added more properties to Item, re-factored code
Renamed Item's OrientationFlag->Orientation to match EXIF name
Added new Item properties for the image metadata: OrientationXMP,
JPEGInterchangeFormat, JPEGInterchangeFormatLength, SubjectArea
Added new helper function TempFilePath to retrieve the temporary path
associated with an Item
Added new helper class OrientationMatrix to hold details of the
transformation matrices per Orientation value and whether the X-Y axes
are flipped
Now using the XMP Orientation value if the IFD one not available
Added placeholder value for whether to delete the thumbnail data or not
(currently set to delete)

Updated error propagation so errors return immediately when they can.
Renamed GetJPEGOrientationFlag to GetMetadata. This function now
operates on an Item. Returning better error
WINCODEC_ERR_UNKNOWNIMAGEFORMAT if the file format is not JPEG. Moved
most of the reads of metadata into this function so there is now need to
re-read it in FixMetadata.
Moved the code to delete the App1/1st IFD block to its own function
DeleteJPEGThumbnailData now operates on an Item. It now requests
sequential read access to the temporary file. It now has more
informative error codes. Added some comments. Can now use MoveFileExW
with MOVEFILE_REPLACE_EXISTING to swap in the temp file because I've
untangled the code so that no open handles are left hanging when you're
trying to deal with the metadata.
FixMetadata now operates on an Item.
CreateDecoderFromStream->CreateDecoderFromFilename, so there is now
longer a need to open async WinRT handles to the temp file.
CreateReorientedTempFileAsync now operates on an Item.