Rotating thumbnails, fixing metadata out-of-place
Added new Item fields for the existence of thumbnail, EXIF's
SubjectLocation
Changed old C-style casts with C++ casts for pointers
Fixed property change bug for PtrSubjectArea
Replaced ".Get()->QueryInterface(IID_PPV_ARGS(" constructs on smart COM
pointers with ".As("
Renamed OrientationMatrix to OrientationHelper as it not only for
matrices.  Added WICBitmapTransformOptions to OrientationHelper. Fixed
bug with _XYFlips not being set for orientation value of 8.
GetMetadata will now not enumerate all metadata if it knows that the app
won't rotate the image automatically, making for faster analysis
Added warning about incompleteness of the DeleteJPEGThumbnailData code
Added new function FixMetadataOutOfPlace that will fix metadata that
cannot be changed in-place. This also allows me to change the thumbnail
using only Microsoft APIs.
Commented out code path to delete the thumbnail
Fixed race condition where the app appeared to hang because the analysed
counter updated before the to-be-rotated counter