Changed getting metadata to be an async function
- this tidies up the access to the file stream
Item's Title now comes directly from the _StorageFile within
The number of images being rotated through the JPEG library can now not
exceed the number of available physical processors. When there are many
images being processed at a time, the JPEG library runs out of memory,
so this will help to alleviate the problem (may still be an issue with
huge JPEG files). This throttling ought to make the application feel
more responsive as the increment of images rotated will go up smoothly
instead of being stuck on a certain amount and then jump straight to
completed.
Consolidated the code to display error if metadata updates fail
Removed duplication of code when incrementing the to-be-rotated counter