Added C JPEG library error handling
to CreateReorientedTempFileAsync so as to error neatly instead of
crashing when the library calls the default error_exit function
Changed orientation helper to be invoked only when an Item's orientation
is not horizontal
Redefined ANSI memory manager function jpeg_mem_available to use size_t
instead of long
Added more comments to CreateReorientedTempFileAsync and added new
argument to use for the maximum RAM the library can consume per
function. Now specifying that the RAM per image should be 90 % of the
max across all images (1 GiB) divided by the number of processors used
Renamed function StorageFileToFilePointer to
StorageFileToFilePointerAsync to be more explicit about the async nature
of it. Added commented code that reads the bytes without using
GetPointerToByteData
Merged two dispatcher calls to set an Item's orientation values