Worked around CopyAndReplaceAsync bug
by reading into the original file stream directly. The bug was that for
large files being worked in parallel Windows threw an exception in
CopyAndReplaceAsync that could not be caught as the method returns a
void task. It appeared as if Windows sometimes just locked the original
files and would not release them
Moved concurrency interruption points to be consistently after the
SUCCEEDED(hr) test