Update for Universal Windows Platform/Windows 10
TODO: Note that the PhoneProductId is most likely wrong. Also, change
logo from 30x30 to 44x44 and 70x70 to 71x71
It is no longer possible to get the user name without asking permission,
so store the app settings globally. This change removes the
asynchronicity of the loading and saving functions of the settings,
hence renamed them. As they are not synchronous, there is no need to set
sensible defaults at launch.
The Settings charm is unusable in UWP, so removed any references to it
at start-up (the source XAMLs for the settings pane are untouched)
Removed Generated Files directory, as it will be re-created if required
from the source project
No longer generating debug information for Release builds
Concurrency::interruption_point is no longer supported, so commented out
the interruption points for now
Changed namespace from concurrency to Concurrency to avoid the redirect